feat(content): move Seasonal Events to /dashboard/content/seasonal-events
Per user feedback — Seasonal Events is content-domain configuration
(feeds the supervisor's seasonal generator stream + the strategist),
not a settings-domain detail. Promoted to a standalone page under the
Content menu group sitting after Calendar.

Also enriches the event card: when a moving holiday has a date mapped
for the current year (Diwali, Eid, Easter, …), the badge now shows
the weekday + month + day (e.g. "Sat · Nov 8") instead of just the
modal month. Static-date events keep the bare-month badge since
the day-of-month isn't carried in the schema.

Implementation:

1. **New page** at `/dashboard/content/seasonal-events`. Self-
   contained — re-imports `useAuth` + the `/v1/dashboard/org` GET
   + the existing PUT path with `taxonomy.seasonalEvents`. Same
   editable-DateRow + clientId pattern from peakhour-b2c #192 + #193
   (no duplicate-year drift, sparse keys preserved, 20-row cap).
2. **Settings page trimmed** — all seasonal-event types (`SeasonalEvent`,
   `DateRow`, `EditableSeasonalEvent`), helpers (`makeClientId`,
   `nextYearForRows`, `attachClientIds`, `MONTH_SHORT_NAMES`), state
   (`editingSeasonal`, `editSeasonalEvents`, `savingSeasonal`),
   `handleSaveSeasonalEvents`, the Card block, and the two
   sub-components (`ReadOnlySeasonalEvents`, `EditableSeasonalEvents`)
   removed. Country edit + sectors + budget + plan stay on settings.
   Unused imports (`CalendarDays`, `Select*`) cleaned up.
3. **Nav** — `{ href: "/dashboard/content/seasonal-events", label:
   "Seasonal Events" }` added to the Content group's subItems after
   Calendar. The IA comment updated to reflect the new ordering.
4. **Date badge enrichment** — new `formatEventDateBadge(event)`
   helper. When `event.dates[currentYear]` exists and parses, it
   returns `"Sat · Nov 8"` (UTC-anchored so the weekday doesn't shift
   by viewer timezone). Otherwise falls back to the modal-month
   abbreviation. Used only in read view — edit mode keeps the month
   dropdown as the editable control.

Settings is no longer the discovery surface for editing the calendar;
owners go to Content → Seasonal Events. The country-code edit on
Settings → Business Details continues to indirectly affect onboarding's
seasonal-pack lookup.

Typecheck + lint clean.
