+++ src/app/dashboard/settings/page.tsx
@@ -29,16 +29,8 @@ import {
   X,
   Check,
   Plus,
-  CalendarDays,
   Globe2,
 } from "lucide-react";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
 
 interface OrgDetails {
   _id: string;
@@ -71,7 +63,6 @@ interface OrgDetails {
   taxonomy?: {
     sectors?: string[];
     generatedAt?: string;
-    seasonalEvents?: SeasonalEvent[];
   } | null;
   location?: {
     country?: string;
@@ -81,102 +72,6 @@ interface OrgDetails {
   };
   createdAt?: string;
 }
-
-interface SeasonalEvent {
-  name: string;
-  /** 1..12 */
-  month: number;
-  relevance?: string;
-  /** Year (YYYY) → ISO date (YYYY-MM-DD). Set by the seed for moving
-   *  holidays; read-only in v0 of this UI. Display only. */
-  dates?: Record<string, string>;
-}
-
-/**
- * Client-side row state during edit. Two shape changes vs the wire
- * `SeasonalEvent`:
- *  - `clientId` — stable React key so removing a middle row doesn't
- *    drift focus + leak input values across rows.
- *  - `dateRows` — per-year dates rendered as an editable array (each
- *    row has its own `rowId` + year + date strings) rather than the
- *    wire `Record<YYYY, YYYY-MM-DD>`. Lets the user freely edit both
- *    year and date without the "editing a record's key" awkwardness.
- *    Converted back to a `Record` at save time.
- *
- * Both fields are stripped before sending to the api — the server's
- * `.strict()` schema would reject them.
- */
-interface DateRow {
-  rowId: string;
-  year: string;
-  date: string;
-}
-
-interface EditableSeasonalEvent {
-  clientId: string;
-  name: string;
-  month: number;
-  relevance?: string;
-  dateRows: DateRow[];
-}
-
-function makeClientId() {
-  // crypto.randomUUID is available in all evergreen browsers + Node 19+.
-  return typeof crypto !== "undefined" && crypto.randomUUID
-    ? crypto.randomUUID()
-    : `e-${Date.now()}-${Math.random().toString(36).slice(2, 9)}`;
-}
-
-/**
- * Pick a sensible default year when the owner clicks "Add date".
- * Returns the year AFTER the largest valid year in the current rows,
- * or the current calendar year if there are no valid rows. Empty
- * string if the next year would exceed a sensible ceiling — the
- * owner can still edit the field after.
- */
-function nextYearForRows(rows: DateRow[]): string {
-  let maxValid = 0;
-  for (const r of rows) {
-    if (/^\d{4}$/.test(r.year)) {
-      const n = parseInt(r.year, 10);
-      if (n > maxValid) maxValid = n;
-    }
-  }
-  const candidate = maxValid > 0 ? maxValid + 1 : new Date().getUTCFullYear();
-  // Cap at a sane ceiling (operator can still edit the field if
-  // they really want 9999). The api regex accepts any 4-digit year.
-  return candidate >= 1000 && candidate <= 9999 ? String(candidate) : "";
-}
-
-function attachClientIds(events: SeasonalEvent[]): EditableSeasonalEvent[] {
-  return events.map((e) => ({
-    clientId: makeClientId(),
-    name: e.name,
-    month: e.month,
-    ...(e.relevance !== undefined ? { relevance: e.relevance } : {}),
-    dateRows: e.dates
-      ? Object.entries(e.dates)
-          .sort(([a], [b]) => a.localeCompare(b))
-          .map(([year, date]) => ({ rowId: makeClientId(), year, date }))
-      : [],
-  }));
-}
-
-const MONTH_SHORT_NAMES = [
-  "",
-  "Jan",
-  "Feb",
-  "Mar",
-  "Apr",
-  "May",
-  "Jun",
-  "Jul",
-  "Aug",
-  "Sep",
-  "Oct",
-  "Nov",
-  "Dec",
-] as const;
 
 export default function SettingsPage() {
   return (
@@ -211,14 +106,6 @@ function SettingsContent() {
   const [editUrl, setEditUrl] = useState("");
   const [editSectors, setEditSectors] = useState<string[]>([]);
   const [editCountry, setEditCountry] = useState("");
-
-  // Seasonal events edit state — kept independent of the Business
-  // Details card so an owner can edit one without touching the other.
-  // `EditableSeasonalEvent` carries a stable `clientId` for React keys
-  // so removing a middle row doesn't drift input focus / values.
-  const [editingSeasonal, setEditingSeasonal] = useState(false);
-  const [savingSeasonal, setSavingSeasonal] = useState(false);
-  const [editSeasonalEvents, setEditSeasonalEvents] = useState<EditableSeasonalEvent[]>([]);
 
   async function handleSaveBusinessDetails() {
     setSaving(true);
@@ -255,70 +142,6 @@ function SettingsContent() {
       else setError("Failed to save. Please try again.");
     } finally {
       setSaving(false);
-    }
-  }
-
-  async function handleSaveSeasonalEvents() {
-    setSavingSeasonal(true);
-    setError("");
-    try {
-      // Strip empty rows the owner left blank; rebuild the per-year
-      // `dates` record from `dateRows` (the editable array form).
-      // The api's SeasonalEventBody is `.strict()` — clientId +
-      // dateRows + rowId MUST stay local and never reach the server.
-      const cleaned = editSeasonalEvents
-        .map((e) => {
-          // Build the dates record from rows that have both a valid
-          // 4-digit year AND a YYYY-MM-DD date. Last-write-wins on
-          // duplicate years. The 20-entry cap is enforced upstream
-          // by disabling the Add-date button at `dateRows.length >= 20`
-          // — combined with the server's `.refine(<=20)` as defense
-          // in depth, no slice is needed here.
-          const dates: Record<string, string> = {};
-          for (const row of e.dateRows) {
-            if (!/^\d{4}$/.test(row.year)) continue;
-            if (!/^\d{4}-\d{2}-\d{2}$/.test(row.date)) continue;
-            dates[row.year] = row.date;
-          }
-          const hasDates = Object.keys(dates).length > 0;
-          return {
-            name: e.name.trim(),
-            month: e.month,
-            ...(e.relevance && e.relevance.trim()
-              ? { relevance: e.relevance.trim() }
-              : {}),
-            ...(hasDates ? { dates } : {}),
-          };
-        })
-        .filter((e) => e.name.length > 0);
-
-      // Guard against accidentally clearing the whole array. If the
-      // original had events and the cleaned list is empty, prompt for
-      // confirmation before persisting.
-      const originalCount = orgDetails?.taxonomy?.seasonalEvents?.length ?? 0;
-      if (originalCount > 0 && cleaned.length === 0) {
-        const confirmed =
-          typeof window !== "undefined" &&
-          window.confirm(
-            `This will remove all ${originalCount} seasonal events from your business. Continue?`,
-          );
-        if (!confirmed) {
-          setSavingSeasonal(false);
-          return;
-        }
-      }
-
-      await api.put("/v1/dashboard/org", {
-        taxonomy: { seasonalEvents: cleaned },
-      });
-      const updated = await api.get<OrgDetails>("/v1/dashboard/org");
-      setOrgDetails(updated);
-      setEditingSeasonal(false);
-    } catch (err) {
-      if (err instanceof ApiError) setError(err.message);
-      else setError("Failed to save seasonal events. Please try again.");
-    } finally {
-      setSavingSeasonal(false);
     }
   }
 
@@ -476,71 +299,6 @@ function SettingsContent() {
           </CardContent>
         </Card>
 
-        {/* Seasonal Events */}
-        <Card>
-          <CardHeader>
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-2">
-                <CalendarDays className="h-5 w-5 text-muted-foreground" />
-                <CardTitle>Seasonal Events</CardTitle>
-              </div>
-              {!editingSeasonal ? (
-                <Button
-                  size="sm"
-                  variant="ghost"
-                  className="gap-1.5 text-muted-foreground"
-                  onClick={() => {
-                    setEditingSeasonal(true);
-                    setEditSeasonalEvents(
-                      attachClientIds(orgDetails?.taxonomy?.seasonalEvents || []),
-                    );
-                  }}
-                >
-                  <Pencil className="h-3.5 w-3.5" />
-                  Edit
-                </Button>
-              ) : (
-                <div className="flex gap-2">
-                  <Button
-                    size="sm"
-                    variant="ghost"
-                    onClick={() => setEditingSeasonal(false)}
-                    disabled={savingSeasonal}
-                  >
-                    <X className="h-3.5 w-3.5" /> Cancel
-                  </Button>
-                  <Button
-                    size="sm"
-                    onClick={handleSaveSeasonalEvents}
-                    disabled={savingSeasonal}
-                    className="gap-1.5"
-                  >
-                    <Check className="h-3.5 w-3.5" />
-                    {savingSeasonal ? "Saving..." : "Save"}
-                  </Button>
-                </div>
-              )}
-            </div>
-            <CardDescription>
-              Cultural, national, and commercial moments the AI uses to plan
-              seasonal content. Onboarding seeds these from your country&apos;s
-              calendar; you can add, edit, or remove events here.
-            </CardDescription>
-          </CardHeader>
-          <CardContent className="space-y-2">
-            {editingSeasonal ? (
-              <EditableSeasonalEvents
-                events={editSeasonalEvents}
-                onChange={setEditSeasonalEvents}
-              />
-            ) : (
-              <ReadOnlySeasonalEvents
-                events={orgDetails?.taxonomy?.seasonalEvents || []}
-              />
-            )}
-          </CardContent>
-        </Card>
-
         {/* Budget & Guardrails */}
         <Card>
           <CardHeader>
@@ -687,326 +445,3 @@ function EditableSectors({ sectors, onChange }: { sectors: string[]; onChange: (
   );
 }
 
-function ReadOnlySeasonalEvents({ events }: { events: SeasonalEvent[] }) {
-  if (events.length === 0) {
-    return (
-      <p className="text-sm text-muted-foreground">
-        No seasonal events yet. Click Edit to add holidays, festivals, or
-        commercial moments relevant to your audience.
-      </p>
-    );
-  }
-  // Sort by month for a calendar-ish read order. Same-month events
-  // preserve their declared sort order.
-  const sorted = [...events].sort((a, b) => a.month - b.month);
-  return (
-    <div className="space-y-1.5">
-      {sorted.map((e, i) => {
-        const movingYears = e.dates ? Object.keys(e.dates).sort() : [];
-        return (
-          <div
-            key={`${e.name}-${i}`}
-            className="rounded-md border bg-card p-2.5 text-sm"
-          >
-            <div className="flex flex-wrap items-baseline gap-2">
-              <span className="font-medium">{e.name}</span>
-              <Badge variant="outline" className="font-mono text-[10px]">
-                {MONTH_SHORT_NAMES[e.month] ?? `m${e.month}`}
-              </Badge>
-              {movingYears.length > 0 ? (
-                <Badge variant="secondary" className="text-[10px]">
-                  moving · {movingYears.length}y mapped
-                </Badge>
-              ) : null}
-            </div>
-            {e.relevance ? (
-              <p className="mt-1 text-xs text-muted-foreground">{e.relevance}</p>
-            ) : null}
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-function EditableSeasonalEvents({
-  events,
-  onChange,
-}: {
-  events: EditableSeasonalEvent[];
-  onChange: (events: EditableSeasonalEvent[]) => void;
-}) {
-  function updateById(
-    clientId: string,
-    patch: Partial<Omit<EditableSeasonalEvent, "clientId" | "dateRows">>,
-  ) {
-    onChange(
-      events.map((e) => (e.clientId === clientId ? { ...e, ...patch } : e)),
-    );
-  }
-  function removeById(clientId: string) {
-    onChange(events.filter((e) => e.clientId !== clientId));
-  }
-  function addBlank() {
-    onChange([
-      ...events,
-      { clientId: makeClientId(), name: "", month: 1, dateRows: [] },
-    ]);
-  }
-  function updateDateRow(
-    clientId: string,
-    rowId: string,
-    patch: Partial<Pick<DateRow, "year" | "date">>,
-  ) {
-    onChange(
-      events.map((e) =>
-        e.clientId === clientId
-          ? {
-              ...e,
-              dateRows: e.dateRows.map((r) =>
-                r.rowId === rowId ? { ...r, ...patch } : r,
-              ),
-            }
-          : e,
-      ),
-    );
-  }
-  function removeDateRow(clientId: string, rowId: string) {
-    onChange(
-      events.map((e) =>
-        e.clientId === clientId
-          ? { ...e, dateRows: e.dateRows.filter((r) => r.rowId !== rowId) }
-          : e,
-      ),
-    );
-  }
-  function addDateRow(clientId: string) {
-    onChange(
-      events.map((e) =>
-        e.clientId === clientId
-          ? {
-              ...e,
-              dateRows: [
-                ...e.dateRows,
-                // Default the new year to "next unmapped year" so common
-                // case "extend Diwali one more year" is one click.
-                {
-                  rowId: makeClientId(),
-                  year: nextYearForRows(e.dateRows),
-                  date: "",
-                },
-              ],
-            }
-          : e,
-      ),
-    );
-  }
-
-  return (
-    <div className="space-y-2">
-      {events.length === 0 ? (
-        <p className="text-xs text-muted-foreground">
-          No events yet. Add your first one below.
-        </p>
-      ) : (
-        events.map((e, i) => {
-          const rowLabel = e.name || `Event ${i + 1}`;
-          const datesAtCap = e.dateRows.length >= 20;
-          // The badge + "N year(s) mapped" copy below refers to UNIQUE
-          // years that will actually persist. Duplicate years collapse
-          // to one entry (last-write-wins at save) so the row count
-          // would otherwise over-count.
-          const uniqueYears = new Set(
-            e.dateRows
-              .filter((r) => /^\d{4}$/.test(r.year))
-              .map((r) => r.year),
-          );
-          const isDuplicateAt = (idx: number, year: string) => {
-            if (!/^\d{4}$/.test(year)) return false;
-            // Walk forward from idx-1 — duplicates flag the LATER row
-            // (the earlier row stays "clean" since it's the one that
-            // would be overwritten at save).
-            for (let j = 0; j < idx; j++) {
-              if (e.dateRows[j].year === year) return true;
-            }
-            return false;
-          };
-          // Stable key via clientId keeps input focus + value bound to
-          // the row that owns them when middle entries are removed —
-          // an index key would visually leave stale text in the input
-          // that "moves up" after a removal.
-          return (
-            <div
-              key={e.clientId}
-              className="rounded-md border bg-card p-2.5 space-y-2"
-            >
-              <div className="flex flex-wrap items-center gap-2">
-                <Input
-                  type="text"
-                  value={e.name}
-                  placeholder="Event name (e.g. Diwali)"
-                  aria-label={`${rowLabel} — name`}
-                  onChange={(ev) =>
-                    updateById(e.clientId, { name: ev.target.value })
-                  }
-                  className="h-8 text-sm flex-1 min-w-48"
-                />
-                <Select
-                  value={String(e.month)}
-                  onValueChange={(v) =>
-                    updateById(e.clientId, { month: parseInt(v, 10) })
-                  }
-                >
-                  <SelectTrigger
-                    className="h-8 w-22"
-                    aria-label={`${rowLabel} — month`}
-                  >
-                    <SelectValue />
-                  </SelectTrigger>
-                  <SelectContent>
-                    {MONTH_SHORT_NAMES.slice(1).map((label, idx) => (
-                      <SelectItem key={label} value={String(idx + 1)}>
-                        {label}
-                      </SelectItem>
-                    ))}
-                  </SelectContent>
-                </Select>
-                {uniqueYears.size > 0 ? (
-                  <Badge variant="secondary" className="text-[10px]">
-                    {uniqueYears.size}y mapped
-                  </Badge>
-                ) : null}
-                <Button
-                  size="sm"
-                  variant="ghost"
-                  onClick={() => removeById(e.clientId)}
-                  className="h-8 w-8 p-0 text-muted-foreground hover:text-destructive"
-                  aria-label={`Remove ${rowLabel}`}
-                >
-                  <X className="h-3.5 w-3.5" />
-                </Button>
-              </div>
-              <Input
-                type="text"
-                value={e.relevance || ""}
-                placeholder="What does this event mean for your audience? (optional)"
-                aria-label={`${rowLabel} — relevance`}
-                onChange={(ev) =>
-                  updateById(e.clientId, { relevance: ev.target.value })
-                }
-                className="h-8 text-sm"
-              />
-              {/* Per-year dates editor — used for lunar / variable
-                  holidays where the Gregorian date shifts year by year
-                  (Diwali, Eid, Easter, etc.). For static-date events
-                  this section starts empty; the owner can opt-in by
-                  clicking Add date if they want to override the modal
-                  month for a specific year. */}
-              <div className="rounded bg-muted/30 px-2 py-1.5 space-y-1">
-                <div className="flex items-center justify-between">
-                  <p className="text-[10px] font-medium text-muted-foreground">
-                    Per-year dates {datesAtCap ? "(max 20)" : "(optional, lunar/variable)"}
-                  </p>
-                  <Button
-                    size="sm"
-                    variant="ghost"
-                    onClick={() => addDateRow(e.clientId)}
-                    disabled={datesAtCap}
-                    className="h-6 text-[10px] gap-1 px-1.5 text-muted-foreground"
-                    aria-label={`Add per-year date to ${rowLabel}`}
-                  >
-                    <Plus className="h-3 w-3" /> Add date
-                  </Button>
-                </div>
-                {e.dateRows.length === 0 ? (
-                  <p className="text-[10px] text-muted-foreground italic">
-                    None — the modal month is the fallback.
-                  </p>
-                ) : (
-                  <div className="space-y-1">
-                    {e.dateRows.map((row, rowIdx) => {
-                      const yearOk = /^\d{4}$/.test(row.year);
-                      const dateOk = /^\d{4}-\d{2}-\d{2}$/.test(row.date);
-                      const isDup = isDuplicateAt(rowIdx, row.year);
-                      const yearInvalid =
-                        (!yearOk && row.year !== "") || isDup;
-                      return (
-                        <div
-                          key={row.rowId}
-                          className="flex items-center gap-1.5 text-[11px]"
-                        >
-                          <Input
-                            type="text"
-                            inputMode="numeric"
-                            maxLength={4}
-                            value={row.year}
-                            placeholder="YYYY"
-                            aria-label={
-                              isDup
-                                ? `${rowLabel} — year ${row.year} duplicates an earlier row (last wins)`
-                                : `${rowLabel} — year for date row`
-                            }
-                            title={
-                              isDup
-                                ? `Year ${row.year} is set twice for this event — only the last entry saves.`
-                                : undefined
-                            }
-                            onChange={(ev) =>
-                              updateDateRow(e.clientId, row.rowId, {
-                                year: ev.target.value,
-                              })
-                            }
-                            className={`h-7 w-16 font-mono text-[11px] ${
-                              yearInvalid ? "border-destructive" : ""
-                            }`}
-                          />
-                          <span className="text-muted-foreground">:</span>
-                          <Input
-                            type="date"
-                            value={row.date}
-                            aria-label={`${rowLabel} — date for year ${row.year || "?"}`}
-                            onChange={(ev) =>
-                              updateDateRow(e.clientId, row.rowId, {
-                                date: ev.target.value,
-                              })
-                            }
-                            className={`h-7 font-mono text-[11px] flex-1 max-w-48 ${
-                              !dateOk && row.date ? "border-destructive" : ""
-                            }`}
-                          />
-                          <Button
-                            size="sm"
-                            variant="ghost"
-                            onClick={() => removeDateRow(e.clientId, row.rowId)}
-                            className="h-7 w-7 p-0 text-muted-foreground hover:text-destructive"
-                            aria-label={`Remove year ${row.year || "row"} from ${rowLabel}`}
-                          >
-                            <X className="h-3 w-3" />
-                          </Button>
-                        </div>
-                      );
-                    })}
-                  </div>
-                )}
-              </div>
-            </div>
-          );
-        })
-      )}
-      <Button
-        size="sm"
-        variant="outline"
-        onClick={addBlank}
-        className="gap-1.5"
-        disabled={events.length >= 64}
-      >
-        <Plus className="h-3.5 w-3.5" /> Add event
-      </Button>
-      {events.length >= 64 ? (
-        <p className="text-[10px] text-muted-foreground">
-          Maximum of 64 events. Remove one to add another.
-        </p>
-      ) : null}
-    </div>
-  );
-}
